HHH-19438 move OutputableType to org.hibernate.type
we're trying to clean up org.hibernate.Query
