+++ hibernate-core/src/main/java/org/hibernate/sql/exec/spi/JdbcCallParameterRegistration.java
@@ -8,7 +8,7 @@ import java.sql.CallableStatement;
 import jakarta.persistence.ParameterMode;
 
 import org.hibernate.engine.spi.SharedSessionContractImplementor;
-import org.hibernate.query.OutputableType;
+import org.hibernate.type.OutputableType;
 import org.hibernate.sql.exec.internal.JdbcCallRefCursorExtractorImpl;
 
 /**
