+++ hibernate-core/src/main/java/org/hibernate/type/ProcedureParameterExtractionAware.java
@@ -4,11 +4,9 @@
  */
 package org.hibernate.type;
 
-import org.hibernate.query.OutputableType;
-
 /**
- * Optional {@link Type} contract for implementations that are aware of how to extract values from
- * store procedure OUT/INOUT parameters.
+ * Optional {@link Type} contract for implementations that are aware of
+ * how to extract values from stored procedure OUT/INOUT parameters.
  *
  * @author Steve Ebersole
  */
